requirements.txt
Flask
Werkzeug
twilio
python-dotenv